remove trampoline class
