+++ Source/FieldSolver/ElectrostaticSolvers/ElectrostaticSolver_fwd.H
@@ -1,6 +1,8 @@
-/* Copyright 2024 Roelof Groenwald, Arianna Formenti, Revathi Jambunathan
+/* Copyright 2024 The WarpX Community
  *
  * This file is part of WarpX.
+ *
+ * Authors: Roelof Groenwald, Arianna Formenti, Revathi Jambunathan
  *
  * License: BSD-3-Clause-LBNL
  */
