+++ Source/FieldSolver/ElectrostaticSolvers/Make.package
@@ -1,7 +1,6 @@
 CEXE_sources += PoissonBoundaryHandler.cpp
 CEXE_sources += LabFrameExplicitES.cpp
 CEXE_sources += RelativisticExplicitES.cpp
-CEXE_sources += ElectrostaticBase.cpp
 CEXE_sources += ElectrostaticSolver.cpp
 
 VPATH_LOCATIONS   += $(WARPX_HOME)/Source/FieldSolver/ElectrostaticSolvers
